dicom_backend/requirements.txt
fastapi
uvicorn
pydicom
numpy
pylibjpeg
pylibjpeg-libjpeg
pylibjpeg-openjpeg
gdcm
scikit-image
opencv-python-headless
Pillow
python-multipart